Separate file for useGallerySelection
