+++ src/main/webapp/ui/src/eln/gallery/index.js
@@ -17,7 +17,7 @@ import useViewportDimensions from "../../util/useViewportDimensions";
 import Alerts from "../../Inventory/components/Alerts";
 import { DisableDragAndDropByDefault } from "../../components/useFileImportDragAndDrop";
 import Analytics from "../../components/Analytics";
-import { GallerySelection } from "./useGalleryActions";
+import { GallerySelection } from "./useGallerySelection";
 
 function WholePage() {
   const [appliedSearchTerm, setAppliedSearchTerm] = React.useState("");
